Refactor _RealmQueryHandler (now QueryHandle)
